# database/config/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:7001",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/feedback/save": {
      "post": {
        "tags": [
          "feedback-controller"
        ],
        "summary": "Save a feedback object",
        "description": "Saves a new feedback and returns the saved feedback.",
        "operationId": "save_6",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackDto"
                }
              }
            }
          }
        }
      }
    },
    "/feedback/save-all": {
      "post": {
        "tags": [
          "feedback-controller"
        ],
        "summary": "Save multiple feedbacks",
        "description": "Saves a list of feedbacks and returns the list of saved feedbacks.",
        "operationId": "save_7",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FeedbackDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drawing/save": {
      "post": {
        "tags": [
          "drawing-controller"
        ],
        "summary": "Save a drawing",
        "description": "Saves a new drawing and returns the saved drawing.",
        "operationId": "save_8",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrawingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DrawingDto"
                }
              }
            }
          }
        }
      }
    },
    "/drawing/save-all": {
      "post": {
        "tags": [
          "drawing-controller"
        ],
        "summary": "Save multiple drawings",
        "description": "Saves a list of drawings and returns the list of saved drawings.",
        "operationId": "save_9",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DrawingDto"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DrawingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/feedback/get/{id}": {
      "get": {
        "tags": [
          "feedback-controller"
        ],
        "summary": "Retrieve a feedback by its ID",
        "description": "Retrieves a feedback based on its ID. Yields HttpStatus.NOT_FOUND if no feedback was found.",
        "operationId": "getFeedbackById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackDto"
                }
              }
            }
          }
        }
      }
    },
    "/feedback/get-for-history/{id}": {
      "get": {
        "tags": [
          "feedback-controller"
        ],
        "summary": "Retrieve feedbacks by their associated history ID",
        "description": "Retrieves a list of all feedbacks associated with a specific history ID.Yields HttpStatus.NOT_FOUND if no feedback was found.",
        "operationId": "getFeedbacksByHistoryId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/feedback/get-all": {
      "get": {
        "tags": [
          "feedback-controller"
        ],
        "summary": "Retrieve all feedbacks",
        "description": "Retrieves a list of all feedbacks.Yields an empty list if no feedbacks were found.",
        "operationId": "getAllFeedbacks",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drawing/get/{id}": {
      "get": {
        "tags": [
          "drawing-controller"
        ],
        "summary": "Retrieve a drawing by its ID",
        "description": "Retrieves a drawing based on its ID. Yields HttpStatus.NOT_FOUND if no drawing was found.",
        "operationId": "getDrawingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrawingDto"
                }
              }
            }
          }
        }
      }
    },
    "/drawing/get-all": {
      "get": {
        "tags": [
          "drawing-controller"
        ],
        "summary": "Retrieve all drawings",
        "description": "Retrieves a list of all drawings. Yields an empty list if no drawings were found.",
        "operationId": "getAllDrawings",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DrawingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/feedback/delete/{id}": {
      "delete": {
        "tags": [
          "feedback-controller"
        ],
        "summary": "Delete a feedback by its ID",
        "description": "Deletes a feedback based on its ID.",
        "operationId": "deleteFeedbackById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/drawing/delete/{id}": {
      "delete": {
        "tags": [
          "drawing-controller"
        ],
        "summary": "Delete a drawing by its ID",
        "description": "Deletes a drawing based on its ID.",
        "operationId": "deleteDrawingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SearchDataDto": {
        "type": "object",
        "properties": {
          "searchdata_id": {
            "type": "integer",
            "format": "int32"
          },
          "drawing_id": {
            "type": "integer",
            "format": "int32"
          },
          "shape": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "material": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "general_tolerances": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "surfaces": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gdts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "threads": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "outer_dimensions": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "search_vector": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          "part_number": {
            "type": "string"
          },
          "ocr_text": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "runtime_text": {
            "type": "string"
          }
        }
      },
      "RuntimeDto": {
        "type": "object",
        "properties": {
          "runtime_id": {
            "type": "integer",
            "format": "int32"
          },
          "drawing_id": {
            "type": "integer",
            "format": "int32"
          },
          "machine": {
            "type": "string"
          },
          "machine_runtime": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "FeedbackDto": {
        "type": "object",
        "properties": {
          "history_id": {
            "type": "integer",
            "format": "int32"
          },
          "drawing_id": {
            "type": "integer",
            "format": "int32"
          },
          "feedback_desc": {
            "type": "string"
          },
          "feedback_value": {
            "type": "integer",
            "format": "int32"
          },
          "feedback_id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "readOnly": true
      },
...
      "DrawingDto": {
        "type": "object",
        "properties": {
          "drawing_id": {
            "type": "integer",
            "format": "int32"
          },
          "original_drawing": {
            "type": "string"
          },
          "runtimes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuntimeDto"
            }
          },
          "searchdata": {
            "$ref": "#/components/schemas/SearchDataDto"
          },
          "feedbacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackDto"
            }
          }
        }
      }
    }
  }
}